net: skip network restoration if its physical devs are missing
When performing network restoration on boot it is important that the
operation configures as many networks as it can and that the
operation does not fail so that vdsm can start.

In case a device breaks down, it is possible that the host had other
networks configured that would still allow remote access after
reboot.

This patch allows this use case by filtering out bond and net
configuration depending on missing physical devices. It also
introduces error level logs to make it very apparent when
troubleshooting.

Bug-Url: https://bugzilla.redhat.com/1113091
Change-Id: I9f1370696f2398fe2c3cd9ad92424b126bcd4b7c
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29313
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29832
Reviewed-by: Sandro Bonazzola <sbonazzo@redhat.com>