HostStats: stop storing list of host ifids and speeds
Current code keeps an almost-up-to-date list of host interface names as
a data member of the HostStatsThread. This information is redundant,
since host ifids are already available in HostSample.interfaces.

With the list of ifids, this patch removes a list of their respective
speed. Each speed is stored, instead with its respective
InterfaceSample.

Change-Id: Iad0605c910564a8a7072484301c51b4b8a325751
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29470
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29645
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29689