signals: Handle signals to non-main threads
This patch addresses an issue of signals being delayed in supervdsm,
in attempt to add zombiereaper to supervdsm, zombiereaper did not
receive signals it was supposed to receive.

When any thread in python receives a signal, python's signal handler
(running in that thread) marks the signal as received and writes down that
the main thread should handle any arrived signals. When main thread
calls signal.pause(), it goes to sleep until it receives a signal (on
main the main thread itself).
If during that time, a signal is received on any other thread, it is not
handled until main thread itself receives any other signal (because of what
the signal handler does)

I've reported the issue here, also with a simple reproducing script:
https://bugzilla.redhat.com/show_bug.cgi?id=1114434

Change-Id: I5dbcd00cec22ef12f2b6253b016dcbd0aa889583
Signed-off-by: Dima Kuznetsov <dkuznets@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29392
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>