introduce connectivity log
Users have been requesting to have access to a simple log file that
tracks host connectivity. This patch adds /var/log/vdsm/connectivity.log
for that purpose. It adds a line for any change of connectivity of
tracked interfaces.

Note that the connectivity change is not implemented with `ip monitor`,
since the latter seems to ignore link speed changes, which are required
to be logged.

Bug-Url: https://bugzilla.redhat.com/1111234
Change-Id: I86dc41223babe43194add7ce778c6567b2bc5823
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29472
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29646
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>