api: Make remove optional
According to the scheama, the "remove" argument in
StoragePool.disconnect is optional, but in API it is defined as
positional argument and requires a value.

This error was probbaly hidden by the fact that engine send False when
using xmlprc. But when using jsonrpc this cause the method call to fail,
which eventually cause host to be stuck in "Unassigned" state in the
engine side. This fix the error on the vdsm side.

Change-Id: Ia6948e4fd1ddefd7e33406a214f0cbfe301c958b
Bug-Url: https://bugzilla.redhat.com/1115033
Relates-to: https://bugzilla.redhat.com/1115044
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29465
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29480