domainMonitor: Refine names
The domain monitor use to have monitoredDomains and poolMonitoredDomains
properties. The word "monitored" in these names is unneeded, as the
class is called DomainMonitor, and it does not keep other lists of
domains.

I replaced the _domains variable to the more specific _monitors, since
this dict keep domain monitors, not domain objects.  This is required
now to prevent confusion, since the new domains property does not return
the dict but its keys.

Finally, when we get a monitor from the _monitor dict, we call it
monitor instead of domainThread. It is true that this object is related
to a domain and it run in a thread, but responsibility of this object is
monitoring.

Change-Id: I141d4a0baf6b72ddfa5075052f7f4b24ce336a1a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29532
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>