upgrade: fix seal persistence for oVirt node
The oVirt node filesystem is not persistent by default. However, it
has a mechanism to persist files and directories. It is necessary
for a mechanism such as the upgrade sealing, that depends on files
surviving reboots, to make sure that the files are persisted.

This bug had a crucial effect on the referenced bug. On boot:
- the upgrade to unified script would run before restore-nets
- Since restore-nets hadn't run, nothing would be on running config
- upgrade to unified would save that empty running config as
  persisted
- restore-nets would then run and restore to a non existing
  configuration.

Change-Id: I7bd945ae9fc2b5687a1f216f67711dc5fa3e41aa
Bug-Url: https://bugzilla.redhat.com/1108571
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29545
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Dima Kuznetsov <dkuznets@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Tested-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29822