virt: Restore BC for block type live snapshots
Commit 02741ce introduced a regression in older hosts running libvirt <
1.2.2 where libvirt rejects the type='block' and source/dev combination.
Eric Blake suggested that this could be worked around by supplying the
old format and new format in the domainshapshot XML document.  This
union would continue to work as before on old hosts but would pick up
the correct block snapshot behavior on newer versions.  Basically, we
still express the disk type but we repeat the newPath inside both a dev
and file attribute.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1115126
Change-Id: I85290f17842ada6c04695bf45ffd1b1108c63d55
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29567
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30228
Reviewed-by: Allon Mureinik <amureini@redhat.com>