netinfo: replace the last dependency on ethtool with netlink
After the last work on the lib/vdsm/netlink package and the extensive
usage of it throughout netinfo, there was little reason to continue
to have two libraries for our network information gathering. We still
have vdsm/network/tc.py and testing dependencies on ethtool that will
soon be gone. When that happens, we can remove it from packaging
requirements.

Change-Id: I32bc77d2af0fd973b31f77003652daa0912edfe4
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29601
Reviewed-by: Dan Kenigsberg <danken@redhat.com>