virt: Change log level for disk stats collection failure
When hotplugging a new disk, samples for the newly added device will not
be immediately available which results in a KeyError Traceback in
vdsm.log.  Since this condition is expected, downgrade the logging level
from exception to warning.  In the case of a hot plugged disk, the
failure resolves once enough samples have been collected.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1112998

Change-Id: If04623778b590f1f76eedbdcb2eda433d0a24514
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29677
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>