utils: Add stopwatch for timing operations
stopwatch is a context manager that make it easy to time block of code
that may take significant amount of time.

Example:

    with stopwatch("Resized multipath map %r" % name):
        _resize_map(name)

Will log a debug message:

    Resized multipath map 'dm-42': 0.12 seconds

If debug log level is not enabled, this contents manager does nothing.
The log can be disabled by disabling the vds.stopwatch logger.

Change-Id: Ie0a60538936dbb4596243abe9d731779fd9efb47
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/29685
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>