connectivity log: report if Engine has been seen recently
Vdsm monitors when it has been accessed by a client (most probably
Engine). If Vdsm finds out that it has not been accessed for the last 30
seconds (twice the Enginei's polling rate), "recent_client:False" is
logged to the connectivity.log. When polling is restored,
"recent_client:True" is reported.

Bug-Url: https://bugzilla.redhat.com/1122010
Change-Id: Ice088421d82f60e59ada6e743aeac2704c44cb7f
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30226
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30523
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30574