mailbox: Remove fatal, uneeded, and racy check
After errors in _checkForMail, the error handling code check if the
_inLock is locked and release it. This check is pointless as
_checkForMail use try finally block to acquire and release this lock.
There is no chance that this lock will not be released when returning
from _checkForMail.

This check cause the mailbox thread to exit when using pthreading 1.3,
which reveal the fact that the locked() method was never implemented by
pthreading. This eliminate the fatal failure.

Change-Id: I46af1988a88a8aab3d9835034d68d25eec7d075d
Bug-Url: https://bugzilla.redhat.com/1117795
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29854
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>