mailbox: Fix unsafe locking
In setMaxHostID, both _inLock and _outLock are acquired and released in
an unsafe way, not ensuring that the locks are released on errors.  Both
locks use now with block. The rest of the file use currently try finally
but I plan to convert these to with in a later patch.

Change-Id: I156ddf23796000340d5f68977602fefb5e1f6722
Relates-to: https://bugzilla.redhat.com/1117795
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29855
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>