iface_sampling: Fix debug logging samples
Currently we have a debug logging entry whenever the tx/rx rates go
over 100%. In this log, we try to help debugging of the issue by
printing the sample value of rx and tx for all the samples in the
AVERAGING_WINDOW (size 5).

The problem comes from the fact that the decision to log is taken
by checking samples hs0 and hs1 iterating over all the devices
existing on the older hs1. However, there is no guarantee that
the devices would exist in hs2, hs3, ..., hs_AVERAGING_WINDOW.

The fix is to simply not include the samples for which the device
is missing.

Change-Id: Ibaa80f5fe069875d8e48deb1783f102b2da8f39d
Bug-Url: https://bugzilla.redhat.com/1116577
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29916
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>