arbitrary_vlan_name: Fix network deletion
Up until now, when the network deletion operations were not being
forced, the assertion that the bridge was clean, i.e., that it did
not have other ports (vnics most likely) would fail because it
expected

    bondname.vlanid
or
    nic.vlanid

when the name could be any string. This patch addresses the issue
by making netinfo give us the exact vlan name.

Change-Id: I715806db38ce621098fe5f364e16e874a5ed01ed
Bug-Url: http://bugzilla.redhat.com/1119182
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/30013
Tested-by: Ido Barkan <ibarkan@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>