Live Merge: Stop scanning volume chains at VM startup
During VM create (or recovery) we are scanning the volume chain of each
disk to catch and handle metadata sync issues.  These can happen if a
block job completed while vdsm was not running or if a block job
completed when the VM was running on another host but the metadata could
not be corrected before it was restarted on this host.

Both of these negative flows are relatively rare but the volume chain
scanning code incurs a startup penalty on every create/recoverVM.  Drop
the scanning and handle these cases differently.  To address the first
scenario, we are going to move away from using the libvirt events and
rely solely on our cached _blockJobs info stored in the VM recovery
file.  For the second case, we are going to cooperate with the engine
and allow it to resubmit the same live merge command when the end state
of the first submission is ambiguous.  Vdsm will handle the resubmit by
ensuring the metadata is synchronized.

Change-Id: I48ba5dd9a70f6daa9bb9da94c2f06296f2b788da
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30045
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>