Live Merge: Ignore libvirt block job events
Since VDSM needs to handle missed block job completion events that may
have been emitted while the daemon was stopped, it is best to not rely
on the events at all and initiate all cleanup operations from the
existing polling function.  Since a volume chain sync can be a
relatively expensive operation involving HSM calls, run it in a separate
thread.  Depending on the current state of the storage, the sync may
need to be retried.

Change-Id: Ibbcdca4c0c0e45e9323ecfef9ce2fce10d8451a5
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30046
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>