sampling: do not produce stats for missing disks
Adding of a new disk using hotplug (hotplugDisk verb)
can trigger an error during the stats reporting (getVmStats verb).
This error is transient, so it was likely got unnoticed so far.

When a disk is hot-plugged, VDSM correctly updates the list
of the drives of a VM.
When stats for such a VM are asked, VDSM iterates over the list of
the drives of a VM, and  looks up for disk samples of each drive.
The stats are collected (by default) every 60s, and VDSM considers
the oldest and the newest samples; so, until the oldest set of
samples collected includes values for the hotplugged disk,
we will have an error.

We have a vulerabilility window up to
(sampling_window * sampling_interval)
in the worst case. With default values this is 2 * 60s = 120s.
After that, everything goes back to normality, and the
stats for the new disk are reported correctly.

This patch fixes this issue by just ignoring stats for disks
which are part of a VM, but for which samples aren't
yet available.

Change-Id: If93572c21ecd47fe0bb2a6c6e13c0555034be881
Bug-Url: https://bugzilla.redhat.com/1119051
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30029
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>