tc: add hfsc class parsing support
In order to add Quality of Service reporting and configuring to vdsm,
it is necessary to be able to parse Hierarchical Fair Service Curves
class printouts as provided by the 'tc' iproute2 tool.

This patch follows in the steps of vdsm/network/tc/qdisc.py by
providing a spec dictionary based parser that simplifies a lot the
verification (a test is provided as well).

Change-Id: Ie10373eafbc9316f13994a5b5fe3f4ccda0554b8
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30132
Reviewed-by: Dan Kenigsberg <danken@redhat.com>