vm: ensure valid Vm._dom before domDependentInit
If we reach the _domDependentInit stage of a VM
creation, for every possible flow, including
recovery, we must guarantee that the _dom
attribute is not None, thus VDSM and libvirt
are connected.

Quite a lot of errors could be ignored on
recovery, but this cannot, ever.

This patch moves the '_dom is not None'
check inside _domDependentInit, and makes
it mandatory for each flow.

Change-Id: Ie2840b066260790d4f3079a9cd34f87a70e96891
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30146
Reviewed-by: Dan Kenigsberg <danken@redhat.com>