Do not update timestamp on API.getVMList
Commit 95302a moved the call to _updateTimestamp from the XMLRPC binding
to the API level. This has the adverse effect of having a local API
call, that does not use networking at all, to update the client
timestamp.

Since MoM calls getVMList periodically, this effectively breaks the
connectivityCheck feature of the setupNetworks and editNetwork API
calls: even if such a call breaks network configuration, Vdsm thinks
that it is still reachable by a client, and maintains the broken
configuration instead of reverting it.

Bug-Url: https://bugzilla.redhat.com/1119024
Change-Id: I9f8d4a6de9f79d91c667e577934fdaa2b762beb1
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30178
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30256
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>