tc: add command line wrappers for qdiscs, classes and filters
Up until now there were ad-hoc wrappers in __init__.py for the cases
in which we needed to perform an operation.

This patch adds to each of the entities modules:
    - cls
    - filter
    - qdisc
some more generic operations:
    - add (except filter, can use replace instead)
    - delete
    - replace
    - show

With these new wrappers, it will be possible to start the
integration of addNetwork/delNetwork with tc to implement the QoS
flow.

Change-Id: Icd45c12b3b2eef5f1c3abec784257ee51e118ac9
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30197
Reviewed-by: Dan Kenigsberg <danken@redhat.com>