QoS: Remove broken libvirt-based host network QoS
Back when the current host network level QoSa was added, it was
thought that libvirt had a way of configuring and honoring the
Quality of Service definitions that we were giving it. Unfortunately
that was an unimplemented part of their api, which forced us to
design a new implementation with different semantics.

This patch removes most of the host based network QoS implementation
from the codebase to prepare vdsm for the following patches which
will use the new tc package to provide outbound host network QoS.

Change-Id: I7b82cc6b656fb6af8caa2940c9ee2d14287f3526
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30198
Reviewed-by: Dan Kenigsberg <danken@redhat.com>