pyroute2: optional configurator added
Pyroute2 was added as new optional configurator. pyroute2 library, which it
requires, is not listed as a dependency because it is not strictly necessary.
If you want to use this configurator, you have to install current library
version (0.2.12) from pip or GitHub https://github.com/svinota/pyroute2. If
someone tries to use it and does not have it installed, it writes an ERROR log
and uses Ifcfg instead.

Necessary changes were made in spec file, debian install file and Makefile
because of the new file network/configurators/pyroute_two.py

It makes sense to have iproute2 and pyroute2 options in config file.
pyroute_two looks ugly. Anyway we can not rename pyroute_two.py too because of
import conflicts with library.

Change-Id: I12a1c2b9c20b3d47f4b255d6a3502feee1cfbbdb
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30474
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>