Stomp heartbeat implementation
We sent heartbeat header with CONNECT frame and check whether we
recieve regular stomp frame or EOL in time specified. We check the
heart beat time everytime we need to write to a channel. If we have't
received heartbeat for longer than specified we invalidate the channel
and we reconnect before writing.

When heartbeat interval is set to vdsm it is reduced by 10% to mitigate
race condition (engine invalidates connection, vdsm sends heartbeat).

This change won't mitigate command level timeout to occur but it will
reduce the probability to be similar to xmlrpc.


Change-Id: Ida5cf89951875ae8b8d2af9cdc29dac77c3e0d72
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>