Changing isconfigured parameter semantics
Currently isconfigured holds boolean and that way the name is misleading as
True doesn't mean the module is actually configured. This patch changes
that and put in isconfigured just the ret value of isconfigured() on the
specific module instead of calling it twice, and later on we use this
parameter to decide when to reconfigure the module or not.

Change-Id: Ifb7f6c234993701d912ecf731b4927c380fd7d13
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30571
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: mooli tayer <mtayer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>