dhclient: fix kill_dhclient for dhcp reconfigure flow
When configuring a device with dhclient, we always ask dhclient to
kill any dhclient daemon that there might be running for the device,
so that the new dhclient that we (or initscripts) launch will not
abort due to it.

Unfortunately, I did not know nor consider that dhclient, when
checking the interface for discovering, checks that it does not have
any IP address configured other than '0.0.0.0'. Thus, in order for
the new dhclient invokation to succeed all the way to daemonization,
it is necessary to flush the ip addresses of the device.

Another mistake it had was assuming that the pidfile would be
found by dhclient -x, but it only knows to look on a path that
is not used by initscripts nor other tools.

Bug-Url: https://bugzilla.redhat.com/1122477
Change-Id: I62e16a24066c7fcaf19dcb89787cd8adcd3cdff3
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30651
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30852