spec: Create default multipath.conf during install
/etc/multipath.conf does not exits after multipath is installed. This
causes multipathd to fail silently when started as vdsmd rquired
service. Later, when vdsm install new configuration and try to reload
multipathd, it fails and panics.

Example multipathd start on RHEL 7:

    [root@reserved-0-247 ~]# systemctl start multipathd
    [root@reserved-0-247 ~]# systemctl status multipathd
    multipathd.service - Device-Mapper Multipath Device Controller
       Loaded: loaded (/usr/lib/systemd/system/multipathd.service; enabled)
       Active: inactive (dead) since Fri 2014-07-25 13:55:48 IDT; 14min ago
               start condition failed at Fri 2014-07-25 14:10:04 IDT; 2s ago
               ConditionPathExists=/etc/multipath.conf was not met

On EL6, multipathd run when configuration is not available, blacklisting
all devices. It log this message to /var/log/message in this case:

    Jul 25 15:25:28 dhcp-1-210 multipathd: /etc/multipath.conf does not
    exist, blacklisting all devices.

This patch uses mpathconf to create a default multipath.conf if the file
does not exists during installation. This configuration blacklist all
devices, simulating the behavior of multipathd on EL6.

Change-Id: I607b4bbf3451612c25bf3da9640b2410314195c1
Bug-Url: https://bugzilla.redhat.com/1120209
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30704
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>