net_configurators: Fix optional configurator importing
With the arrival of the first truly optional (its dependencies do not
need to be installed) configurator, i.e., pyroute_two, a flaw in my
getConfiguratorClass was exposed.

With this patch, all configurators should explicitly define a
    is_available()
method that returns True if it is required and False if it is
optional and some of its dependencies are missing. Additionally,
all the extra dependency importing should not raise.

Change-Id: Iff38f5e334167e34ba379be164e690f2888f5fea
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30924
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>