vm: remove guest agent channels tunables
We had tunables to disable vm/ga channels creation from a long time,
but it was just on debug purposes and these settings should never
be used in production.

Moreover, we acttually rely on the channel being here (see
RHBZ#1126887).

Since it is now possible the achive the same results through hooks,
it is time to drop these tunables.

Change-Id: I4774441d7bd6dcdbb58e38886eb522851d4af37c
Relates-To: https://bugzilla.redhat.com/1126887
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/31245
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>