Live Merge: Update path using libvirt supplied value
When libvirt pivots the active layer after a live merge, the live XML is
updated to reflect the path of the new active layer.  The new path is
copied from the backingStore pointer of the old active layer.  Since
that is a relative path, libvirt makes it absolute but it will not
remove the '../' path component since that is not always correct (even
though it would be correct for our environment).  When syncing the
volume chain, we must use the same path as libvirt does, otherwise
future calls to libvirt may fail.

Change-Id: Ie341333ad934645c0fe4f2fd31733c459d306d9a
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1109920
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/31365
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>