vdsm-tool: changing root checking in configurator.
Checking for root as we currently do in configure of libvirt and
sanlock is not enough. Currently this will fail during the
isConfigured() check since it does not have premissions to the
files it attempts to check:
"OSError: No such file or directory: /etc/libvirt/libvirtd.conf"

It is easy to see that this check is needed for all the exposed
methods of configurator. I'm suggesting to do it in the exposed
verbs in a uniform manner.

Change-Id: I52967e30f677e4537b83c2db442963e3eadecb55
Signed-off-by: Mooli Tayer <mtayer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/31293
Reviewed-by: Dima Kuznetsov <dkuznets@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>