Live Merge: Return more info from _driveGetActualVolumeChain
In followup patches, we want to return more information about each path
component from _driveGetActualVolumeChain.  We need the actual path as
given in the domain XML in order to properly update our metadata after
an active layer merge.  We also want to use this function to return high
write watermark information for block volumes.  This patch just converts
the format returned by _driveGetActualVolumeChain from a simple list to
a list of named tuples.

Change-Id: I01d142a68903048accc95eb04d9930c326965db0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1109920
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/31364
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>