vm: xml: add a method do append custom device XML
add a method to vmxml.Domain to append the XML representation
of a device. This is used to update the XML after the hooks are
been run.

This is needed to achieve better isolation among
vm.Vm and vmxml.Domain, and brings us one tiny step closer
to the isolation of all XML-related processing into vdsm/virt/vmxml.

Change-Id: Ic1c2842e31ade6e1a3c5bb967fcb02d91b81e26b
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/31381
Reviewed-by: Dan Kenigsberg <danken@redhat.com>