tc: Integrate Host QoS with network deletion via setupNetworks
This patch makes it possible to leverage the network.tc package
to remove outbound Host network Quality of Service.

When deleting the last shaped network of a device the qdiscs will
be deleted so that they go back to being the kernel default (just
a single pfifo_fast, at the moment).

Change-Id: I28a3676af1553052e7cdea515faa29e847b7d3c0
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/31440
Reviewed-by: Dan Kenigsberg <danken@redhat.com>